+++ pkg/dtos/search_component_output.go
@@ -12,9 +12,11 @@ type ComponentsSearchOutput struct {
 }
 
 type ComponentSearchOutput struct {
-	Component string `json:"component"`
-	Purl      string `json:"purl"`
-	Url       string `json:"url"`
+	Name      string `json:"name"`
+	Component string `json:"component"` // Deprecated. Component and name fields will contain the same data until
+	// the component field is removed
+	Purl string `json:"purl"`
+	Url  string `json:"url"`
 }
 
 func ExportComponentSearchOutput(s *zap.SugaredLogger, output ComponentsSearchOutput) ([]byte, error) {