+++ pkg/usecase/component.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	se "scanoss.com/components/pkg/errors"
 
 	"github.com/jmoiron/sqlx"
 	"github.com/scanoss/go-grpc-helper/pkg/grpc/database"
@@ -67,11 +68,14 @@ func (c ComponentUseCase) SearchComponents(request dtos.ComponentSearchInput) (d
 
 	for _, component := range searchResults {
 		var componentSearchResult dtos.ComponentSearchOutput
-		componentSearchResult.Component = component.Component
+		componentSearchResult.Name = component.Component
+		componentSearchResult.Component = component.Component // Deprecated. Remove in future versions
 		componentSearchResult.Purl = "pkg:" + component.PurlType + "/" + component.PurlName
 		componentSearchResult.Url = component.Url
-
 		componentsSearchResults = append(componentsSearchResults, componentSearchResult)
+	}
+	if len(componentsSearchResults) == 0 {
+		return dtos.ComponentsSearchOutput{}, se.NewNotFoundError("No components found matching the search criteria")
 	}
 	return dtos.ComponentsSearchOutput{Components: componentsSearchResults}, nil
 }
@@ -106,6 +110,7 @@ func (c ComponentUseCase) GetComponentVersions(request dtos.ComponentVersionsInp
 	var output dtos.ComponentOutput
 	output.Purl = request.Purl
 	if len(allUrls) > 0 {
+		output.Name = allUrls[0].Component
 		output.Url = projectURL
 		output.Component = allUrls[0].Component
 		output.Versions = []dtos.ComponentVersion{}
@@ -130,6 +135,9 @@ func (c ComponentUseCase) GetComponentVersions(request dtos.ComponentVersionsInp
 			version.Licenses = append(version.Licenses, license)
 			output.Versions = append(output.Versions, version)
 		}
+	}
+	if output.Name == "" || output.Purl == "" {
+		return dtos.ComponentVersionsOutput{}, se.NewNotFoundError(fmt.Sprintf("purl: '%v' not found", request.Purl))
 	}
 	return dtos.ComponentVersionsOutput{Component: output}, nil
 }