+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Upcoming changes...
+- 
+## [0.6.0] - 2025-09-18
+### Added
+- Added `name` field to component search and version response DTOs
+### Changed
+- Updated component DTOs to use `name` field instead of `component` field
+- Upgraded `github.com/scanoss/papi` to v0.21.0
+### Deprecated
+- Deprecated `component` field in ComponentOutput and ComponentSearchOutput DTOs (use `name` instead)
 
 ## [0.5.0] - 2025-09-04
 ### Changed
@@ -19,5 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - ?
 
+[0.6.0]: https://github.com/scanoss/components/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/scanoss/components/compare/v0.4.0...v0.5.0
 [0.0.1]: https://github.com/scanoss/components/compare/v0.0.0...v0.0.1